+++ flink-runtime/src/main/java/org/apache/flink/runtime/operators/AllReduceDriver.java
@@ -19,14 +19,14 @@
 
 package org.apache.flink.runtime.operators;
 
-import eu.stratosphere.api.common.operators.base.ReduceOperatorBase;
-import eu.stratosphere.util.InstantiationUtil;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.flink.api.common.functions.GenericReduce;
+import org.apache.flink.api.common.operators.base.ReduceOperatorBase;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
 import org.apache.flink.api.common.typeutils.TypeSerializerFactory;
 import org.apache.flink.runtime.operators.util.TaskConfig;
+import org.apache.flink.util.InstantiationUtil;
 import org.apache.flink.util.MutableObjectIterator;
 
 /**
